move specialized align functions to ListK (#1831)
