+++ modules/core/arrow-core/src/main/kotlin/arrow/core/extensions/listk.kt
@@ -11,14 +11,10 @@ import arrow.core.Option
 import arrow.core.SequenceK
 import arrow.core.Tuple2
 import arrow.core.extensions.listk.eq.eq
-import arrow.core.extensions.listk.functorFilter.flattenOption
 import arrow.core.extensions.listk.monad.monad
-import arrow.core.extensions.listk.semialign.semialign
 import arrow.core.extensions.listk.semigroup.plus
 import arrow.core.fix
 import arrow.core.k
-import arrow.core.leftIor
-import arrow.core.rightIor
 import arrow.core.toT
 import arrow.extension
 import arrow.typeclasses.Align
@@ -314,52 +310,8 @@ interface ListKSemialign : Semialign<ForListK>, ListKFunctor {
   override fun <A, B> align(
     a: Kind<ForListK, A>,
     b: Kind<ForListK, B>
-  ): Kind<ForListK, Ior<A, B>> = alignRec(a.fix(), b.fix()).k()
-
-  private fun <X, Y> alignRec(ls: List<X>, rs: List<Y>): List<Ior<X, Y>> = when {
-    ls.isEmpty() -> rs.map { it.rightIor() }
-    rs.isEmpty() -> ls.map { it.leftIor() }
-    else -> listOf(Ior.Both(ls.first(), rs.first())).listPlus(alignRec(ls.drop(1), rs.drop(1)))
-  }
+  ): Kind<ForListK, Ior<A, B>> = ListK.align(a.fix(), b.fix())
 }
-
-/**
- * Left-padded zipWith.
- */
-fun <A, B, C> ListK<A>.lpadZipWith(
-  other: ListK<B>,
-  fab: (Option<A>, B) -> C
-): ListK<C> =
-  ListK.semialign().run {
-    this@lpadZipWith.padZipWith(other) { a, b -> b.map { fab(a, it) } }.flattenOption()
-  }
-
-/**
- * Left-padded zip.
- */
-fun <A, B> ListK<A>.lpadZip(
-  other: ListK<B>
-): ListK<Tuple2<Option<A>, B>> =
-  this.lpadZipWith(other) { a, b -> a toT b }
-
-/**
- * Right-padded zipWith.
- */
-fun <A, B, C> ListK<A>.rpadZipWith(
-  other: ListK<B>,
-  fa: (A, Option<B>) -> C
-): ListK<C> =
-  other.lpadZipWith(this) { a, b -> fa(b, a) }
-
-/**
- * Right-padded zip.
- */
-fun <A, B> ListK<A>.rpadZip(
-  other: ListK<B>
-): ListK<Tuple2<A, Option<B>>> =
-  this.rpadZipWith(other) { a, b ->
-    a toT b
-  }
 
 @extension
 interface ListKAlign : Align<ForListK>, ListKSemialign {
